Move BMI wrapper into pysb submodule
