+++ indra/assemblers/pysb/bmi_wrapper.py
