+++ indra/tests/test_bmi_wrapper.py
@@ -1,6 +1,6 @@
 from indra.statements import *
 from indra.assemblers import PysbAssembler
-from indra.assemblers.bmi_wrapper import BMIModel
+from indra.assemblers.pysb.bmi_wrapper import BMIModel
 
 stmts = [Influence(Concept('rainfall'), Concept('flood')),
          Influence(Concept('flood'), Concept('displacement'))]
